requirements.txt
diffusers==0.14.0
numpy==1.22.4
Pillow==9.5.0
scikit_learn==1.2.2
streamlit==1.20.0
tensorflow==2.8.0
torch==1.12.1
transformers==4.27.4
